backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Markdown==3.4.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==6.0.0
gunicorn==20.1.0
psycopg2-binary
python-dotenv

